requirements.txt
discord.py==2.3.2
firebase-admin==6.5.0
feedparser==6.0.10
aiohttp==3.9.1
python-dotenv==1.0.0
